requirements.txt
opencv-python-headless==4.8.0.76
mediapipe==0.10.8
numpy>=1.24.3
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
requests==2.31.0
pyrebase4==4.7.1
gunicorn==21.2.0